refactor: cleanup desktop menus
