+++ packages/desktop/components/buttons/index.js
@@ -1,2 +1,1 @@
-export * from './menu-buttons'
 export * from './popup-buttons'
