--- packages/desktop/components/buttons/menu-buttons/index.js
@@ -1,2 +1,0 @@
-export { default as AccountActionsButton } from './AccountActionsButton.svelte'
-export { default as TokenListMenuButton } from './TokenListMenuButton.svelte'
