+++ packages/desktop/components/index.js
@@ -1,7 +1,7 @@
 export * from './buttons'
 export * from './drawers'
 export * from './filter'
-export * from './menu-buttons'
+export * from './menus'
 export * from './modals'
 export * from './panes'
 export * from './popups'
