--- packages/desktop/components/menu-buttons/index.js
@@ -1,1 +1,0 @@
-export { default as ProposalDetailsButton } from './ProposalDetailsButton.svelte'
