+++ packages/desktop/components/menus/ProfileActionsMenu.svelte
