+++ packages/desktop/components/menus/index.js
@@ -1,0 +1,4 @@
+export { default as AccountActionsMenu } from './AccountActionsMenu.svelte'
+export { default as ProfileActionsMenu } from './ProfileActionsMenu.svelte'
+export { default as ProposalDetailsMenu } from './ProposalDetailsMenu.svelte'
+export { default as TokenListMenu } from './TokenListMenu.svelte'
